DrawnRune tags
Tracing and invoking runes stores the drawn rune as a tag to the CombatAction. Tracing a rune will start by storing the Rune, then will store the DrawnRune after it's created.

Damage dealt by runes now tends to store the source DrawnRune as a tag as well, resulting in it being feasible to tell if a damage event's action comes from a specific drawn rune.